NavBar refactor
